Added in enemies that can move and shoot.
Enemies can now move around. They have 2 states, roam and attack. In
their roam state, they move to a random position around them. In the
attack state, which is activated if they are near to the player, they
pursue and shoot at the player. Both enemies and the player are able to
die.